Rename RejectedOffersTable to DeclinedOffersTable
