+++ plugins/services/src/js/components/DeclinedOffersTable.js
